+++ src/ir_Dish.hpp
@@ -38,7 +38,7 @@ void IRsend::sendDISH(unsigned long data, int nbits) {
     mark(DISH_HEADER_MARK);
     space(DISH_HEADER_SPACE);
 
-    sendPulseDistanceWidthData(DISH_BIT_MARK, DISH_ONE_SPACE, DISH_BIT_MARK, DISH_ZERO_SPACE, data, nbits, PROTOCOL_IS_MSB_FIRST);
+    sendPulseDistanceWidthData(DISH_BIT_MARK, DISH_ONE_SPACE, DISH_BIT_MARK, DISH_ZERO_SPACE, data, nbits, PROTOCOL_IS_MSB_FIRST, SEND_NO_STOP_BIT);
     mark(DISH_HEADER_MARK); //added 26th March 2016, by AnalysIR ( https://www.AnalysIR.com )
     IrReceiver.restartAfterSend();
 }